Hoist project detection/setup into 'gcloud._helpers'.
Second bullet for #741.
