+++ gcloud/storage/api.py
@@ -19,8 +19,8 @@ rather than via Connection.
 """
 
 from gcloud.exceptions import NotFound
+from gcloud._helpers import get_default_project
 from gcloud.storage._implicit_environ import get_default_connection
-from gcloud.storage._implicit_environ import get_default_project
 from gcloud.storage.bucket import Bucket
 from gcloud.storage.iterator import Iterator
