+++ regression/storage.py
@@ -19,15 +19,15 @@ import unittest2
 
 from gcloud import exceptions
 from gcloud import storage
+from gcloud import _helpers
 from gcloud.storage._helpers import _base64_md5hash
-from gcloud.storage import _implicit_environ
 from gcloud.storage.batch import Batch
 
 
 HTTP = httplib2.Http()
 SHARED_BUCKETS = {}
 
-_implicit_environ._PROJECT_ENV_VAR_NAME = 'GCLOUD_TESTS_PROJECT_ID'
+_helpers._PROJECT_ENV_VAR_NAME = 'GCLOUD_TESTS_PROJECT_ID'
 storage.set_defaults()
